Add ScrollArea component to design system
