--- app/(navigation)/prompts/components/ScrollArea.module.css
@@ -1,51 +1,0 @@
-.root {
-  overflow: hidden;
-  width: 100%;
-  height: 100%;
-  border-radius: 4px;
-  --scrollbar-size: 10px;
-}
-
-.viewport {
-  width: 100%;
-  height: 100%;
-  border-radius: inherit;
-}
-
-.scrollbar {
-  display: flex;
-  padding: 3px;
-  /* disable browser handling of all panning and zooming gestures on touch devices */
-  touch-action: none;
-  transition: background 160ms ease-out;
-  /* ensures no selection */
-  user-select: none;
-
-  &[data-orientation="vertical"] {
-    width: var(--scrollbar-size);
-  }
-  &[data-orientation="horizontal"] {
-    height: var(--scrollbar-size);
-    flex-direction: column;
-  }
-}
-
-.thumb {
-  position: relative;
-  flex: 1;
-  border-radius: var(--scrollbar-size);
-  background: rgba(255, 255, 255, 0.1);
-
-  /* increase target size for touch devices https://www.w3.org/WAI/WCAG21/Understanding/target-size.html */
-  &::before {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    width: 100%;
-    min-width: 44px;
-    height: 100%;
-    min-height: 44px;
-    content: "";
-    transform: translate(-50%, -50%);
-  }
-}
